requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
google-generativeai
asyncpg
readability-lxml
